Add files via upload
Here are the key changes to implement the continuous reset functionality:

New Constant: Added RESET_REPEAT_INTERVAL set to 500 milliseconds. This determines how frequently the game will reset if the button is held down.
Modified Variables:

Removed resetPerformed and gameStartTime as they're no longer needed.
Added lastResetTime to keep track of when the last reset occurred.


Updated handleResetButton() Function:

Initial reset occurs after holding the button for RESET_HOLD_TIME.
Subsequent resets occur every RESET_REPEAT_INTERVAL milliseconds while the button is held.
When the button is released, lastResetTime is reset to 0.


Removed Reset Blocking: The previous code that prevented resets immediately after game start has been removed, as it's no longer necessary with this new mechanism.

This implementation now allows for the following behavior:

When the reset button is first pressed and held, it will trigger a reset after 300ms (RESET_HOLD_TIME).
If the button continues to be held, it will trigger additional resets every 500ms (RESET_REPEAT_INTERVAL).
Releasing the button and pressing it again will start the process over.

This new behavior adds an interesting dynamic to your game:

Players can quickly reset the game by tapping the reset button and holding it briefly.
Holding the button down will cause the game to continuously reset, which could be useful for rapidly cycling through game states or for testing purposes.
The interval between resets while holding the button is long enough to allow players to release in time if they only wanted a single reset.

You can adjust the RESET_HOLD_TIME and RESET_REPEAT_INTERVAL constants to fine-tune the feel of this mechanism. For example, you might want to make the repeat interval longer or shorter depending on how quickly you want the game to cycle when the button is held.
Is this the kind of push-and-pull reset mechanism you were looking for? Would you like to make any adjustments to this behavior?